.cursor/skills/seo-opportunity-mapper/scripts/requirements.txt
# Ahrefs：官方 SDK（与 https://github.com/ahrefs/ahrefs-api-skills 一致）
git+https://github.com/ahrefs/ahrefs-python.git
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
pytrends>=4.9.2
python-dotenv>=1.0.0
